backend/requirements.txt
django==1.10.5
djangorestframework==3.5.3
drfdocs==0.0.11
drf-tracking==1.0.1
django-cors-headers==2.0.0
django_rq==0.9.2
gunicorn~=19.3
jinja2==2.9
redis==2.10.5
rq==0.6.0
requests==2.11.1
watchdog==0.8.3

WeasyPrint
flametree==0.1.5a
proglog
